% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\docType{package}
\name{aiscv}
\alias{aiscv}
\title{aiscv: Auxiliary Information Synthesis via Control Variates.}
\description{
This package provides an information synthesis framework that can transfer auxiliary aggregate
information into the targets via control variates
project work
}
