% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{HPH.CE}
\alias{HPH.CE}
\title{High-Dimensional Cox proportional hazards model with auxiliary information
in the type of coefficients from external submodels}
\usage{
HPH.CE(
  yobs,
  delta,
  X,
  auxinfo = list(aux = NULL, M = NULL, hetero = TRUE),
  iindex = NULL,
  dsplit = list(do = TRUE, times = 10, combine = "median"),
  control = list(eps = 1e-06, maxit = 10000)
)
}
\arguments{
\item{yobs}{a numeric vector that indicated the observed survival times.}

\item{delta}{a numeric vector that stores the right-censoring indicators
(1 = event of interest happens, and 0 = censoring).}

\item{X}{a matrix that includes all interested covariates.}

\item{auxinfo}{a list used to store information about the inputted auxiliary
information. Specifically, it contains the following elements:
\describe{
\item{\code{aux}}{ a list that includes information about historical
aggregated statistics. It is a list of lists, and each sub-list represents
auxiliary information from the same source study. In each source study,
it contains the following elements:
\describe{
\item{\code{ce}}{ auxiliary regression coefficients.}
\item{\code{ceV}}{ variance-covariance matrix of \code{ce}.}
\item{\code{idx}}{ indices of this submodel's covariates in \code{X}.}
\item{\code{M}}{ an integer that indicate the sample size of the external study.}
} }
\item{\code{hetero}}{ a logical value. If it is \code{TRUE}, the penalization
will be applied to identify the potential heterogeneous auxiliary information
and make a refinement to the final estimator.}
}}

\item{iindex}{indices of our interested coefficients for covariates in \code{X}.
The default is \code{NULL} and all coefficients will be displayed.}

\item{dsplit}{a list used to control the process of data splitting. Specifically,
it contains the following elements:
\describe{
\item{\code{do}}{ a logical that indicate whether the data splitting procedure
should be used. The default value if \code{TRUE}.}
\item{\code{times}}{ an integer that denotes the times of data splitting. The
default value if \code{10}.}
\item{\code{combine}}{ a character that specifies the method of combining
different estimates in each data splitting procedure. It can be either
\code{"median"} (the default) and \code{"mean"}.}
}}

\item{control}{indicates more detailed control of the underlying model fitting
procedures. It is a list of the following three arguments:
\describe{
\item{\code{maxit}}{ a positive integer that denotes the maximum iteration
number in optimization. The default value is \code{10000}.}
\item{\code{eps}}{ a positive small numeric value that denotes the tolerance
for convergence. The default value is \code{1e-6}.}
}}
}
\value{
A list of fitted results is returned.
Within this outputted list, the following elements can be found:
\describe{
\item{\code{coefficient}}{ estimation and inference of regression coefficients.}
\item{\code{tau}}{ estimated degrees of population discrepency.}
}
}
\description{
Fit the High-dimensional Cox proportional hazards (HPH) model with
auxiliary information in the type of coefficients from external submodels (CE).
}
