% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PH.CE}
\alias{PH.CE}
\title{Cox proportional hazards model with auxiliary information in the type
of coefficients from external submodels}
\usage{
PH.CE(
  yobs,
  delta,
  X,
  auxinfo = list(aux = NULL, hetero = TRUE),
  iindex = NULL,
  control = list(maxit = 10000, eps = 1e-06, penalty = list(do = FALSE, inv = TRUE))
)
}
\arguments{
\item{yobs}{a numeric vector that indicated the observed survival times.}

\item{delta}{a numeric vector that stores the right-censoring indicators
(1 = event of interest happens, and 0 = censoring).}

\item{X}{a matrix that includes all interested covariates.}

\item{auxinfo}{a list used to store information about the inputted auxiliary
information. Specifically, it contains the following elements:
\describe{
\item{\code{aux}}{ a list that includes information about historical
aggregated statistics. It is a list of lists, and each sub-list represents
auxiliary information from the same source study. In each source study,
it contains the following elements:
\describe{
\item{\code{ce}}{ auxiliary regression coefficients.}
\item{\code{ceV}}{ variance-covariance matrix of \code{ce}.}
\item{\code{idx}}{ indices of this submodel's covariates in \code{X}.}
\item{\code{M}}{ an integer that indicate the sample size of the external study.}
} }
\item{\code{hetero}}{ a logical value. If it is \code{TRUE}, the penalization
will be applied to identify the potential heterogeneous auxiliary information
and make a refinement to the final estimator.}
}}

\item{iindex}{indices of our interested coefficients for covariates in \code{X}.
The default is \code{NULL} and all coefficients will be displayed.}

\item{control}{indicates more detailed control of the underlying model fitting
procedures. It is a list of the following three arguments:
\describe{
\item{\code{maxit}}{ a positive integer that denotes the maximum iteration
number in optimization. The default value is \code{10000}.}
\item{\code{eps}}{ a positive small numeric value that denotes the tolerance
for convergence. The default value is \code{1e-6}.}
\item{\code{penalty}}{ a list that contains information about the decorrelation steps.
The argument \code{do} indicates whether the penalization should be used,
and the argument \code{inv} controls whether the penalization in the decorrelation
setup should be used.}
}}
}
\value{
A list of fitted results is returned.
Within this outputted list, the following elements can be found:
\describe{
\item{\code{coefficient}}{ estimation and inference of regression coefficients.}
\item{\code{tau}}{ estimated degrees of population discrepency.}
}
}
\description{
Fit the Cox proportional hazards (PH) model with auxiliary information
in the type of coefficients from external submodels (CE).
}
