% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Functions.R
\name{PH.SP}
\alias{PH.SP}
\title{Cox proportional hazards model with auxiliary information in the type
of subgroup survival probabilities at multiple time points}
\usage{
PH.SP(
  yobs,
  delta,
  X,
  auxinfo = list(aux = NULL, M = NULL, hetero = TRUE),
  iindex = NULL,
  control = list(maxit = 10000, eps = 1e-06, penalty = list(do = FALSE, inv = TRUE))
)
}
\arguments{
\item{yobs}{a numeric vector that indicated the observed survival times.}

\item{delta}{a numeric vector that stores the right-censoring indicators
(1 = event of interest happens, and 0 = censoring).}

\item{X}{a matrix that includes all interested covariates.}

\item{auxinfo}{a list used to store information about the inputted auxiliary
information. Specifically, it contains the following elements:
\describe{
\item{\code{aux}}{ a list that includes information about historical
aggregated statistics. It is a list of lists, and each sub-list represents
auxiliary information from the same source study. In each source study,
it contains several time points and each time point contains the following
three elements:
\describe{
\item{\code{tstar}}{ the time point that the auxiliary information
was calculated at.}
\item{\code{sprob}}{ auxiliary subgroup survival rates for each subgroup
at the current time point.}
\item{\code{gfunc}}{ a function used to identify the subgroup.}
} }
\item{\code{M}}{ a vector of integers that indicate the sample sizes of
all external studies. Note that the dimension of this vector should equal
to the number of studies in \code{aux} and the order should be consistent.}
\item{\code{hetero}}{ a logical value. If it is \code{TRUE}, the penalization
will be applied to identify the potential heterogeneous auxiliary subgroup
survival rates and make a refinement to the final estimator.}
}}

\item{iindex}{indices of our interested coefficients for covariates in \code{X}.
The default is \code{NULL} and all coefficients will be displayed.}

\item{control}{indicates more detailed control of the underlying model fitting
procedures. It is a list of the following three arguments:
\describe{
\item{\code{maxit}}{ a positive integer that denotes the maximum iteration
number in optimization. The default value is \code{10000}.}
\item{\code{eps}}{ a positive small numeric value that denotes the tolerance
for convergence. The default value is \code{1e-6}.}
\item{\code{penalty}}{ a list that contains information about the decorrelation steps.
The argument \code{do} indicates whether the penalization should be used,
and the argument \code{inv} controls whether the penalization in the decorrelation
setup should be used.}
}}
}
\value{
A list of fitted results is returned.
Within this outputted list, the following elements can be found:
\describe{
\item{\code{coefficient}}{ estimation and inference of regression coefficients.}
\item{\code{tau}}{ estimated degrees of population discrepency.}
}
}
\description{
Fit the Cox proportional hazards (PH) model with auxiliary information
in the type of subgroup survival probabilities (SP) at multiple time points.
}
